move proxy-related code in component/server into proxy_server
