+++ components/proxy_server/Cargo.toml
@@ -1,0 +1,96 @@
+[package]
+name = "proxy_server"
+version = "0.0.1"
+license = "Apache-2.0"
+edition = "2018"
+publish = false
+
+[features]
+tcmalloc = ["tikv/tcmalloc"]
+jemalloc = ["tikv/jemalloc"]
+mimalloc = ["tikv/mimalloc"]
+snmalloc = ["tikv/snmalloc"]
+portable = ["tikv/portable"]
+sse = ["tikv/sse"]
+mem-profiling = ["tikv/mem-profiling"]
+failpoints = ["tikv/failpoints"]
+bcc-iosnoop = ["tikv/bcc-iosnoop"]
+
+cloud-aws = ["encryption_export/cloud-aws"]
+cloud-gcp = ["encryption_export/cloud-gcp"]
+cloud-azure = ["encryption_export/cloud-azure"]
+test-engine-kv-rocksdb = [
+  "tikv/test-engine-kv-rocksdb"
+]
+test-engine-raft-raft-engine = [
+  "tikv/test-engine-raft-raft-engine"
+]
+test-engines-rocksdb = [
+  "tikv/test-engines-rocksdb",
+]
+test-engines-panic = [
+  "tikv/test-engines-panic",
+]
+nortcheck = ["engine_rocks/nortcheck"]
+backup-stream-debug = ["backup-stream/backup-stream-debug"]
+
+[dependencies]
+api_version = { path = "../api_version" }
+backup = { path = "../backup", default-features = false }
+backup-stream = { path = "../backup-stream", default-features = false }
+causal_ts = { path = "../causal_ts" }
+cdc = { path = "../cdc", default-features = false }
+chrono = "0.4"
+clap = "2.32"
+collections = { path = "../collections" }
+concurrency_manager = { path = "../concurrency_manager", default-features = false }
+crossbeam = "0.8"
+encryption = { path = "../encryption", default-features = false }
+encryption_export = { path = "../encryption/export", default-features = false }
+engine_rocks = { path = "../engine_rocks", default-features = false }
+engine_rocks_helper = { path = "../engine_rocks_helper" }
+engine_traits = { path = "../engine_traits", default-features = false }
+error_code = { path = "../error_code", default-features = false }
+file_system = { path = "../file_system", default-features = false }
+fs2 = "0.4"
+futures = "0.3"
+grpcio = { version = "0.10", default-features = false, features = ["openssl-vendored"] }
+grpcio-health = { version = "0.10", default-features = false, features = ["protobuf-codec"] }
+hex = "0.4"
+
+itertools = "0.10"
+keys = { path = "../keys", default-features = false }
+kvproto = { git = "https://github.com/pingcap/kvproto.git" }
+libc = "0.2"
+log = { version = "0.4", features = ["max_level_trace", "release_max_level_debug"] }
+log_wrappers = { path = "../log_wrappers" }
+nix = "0.23"
+online_config = { path = "../online_config" }
+pd_client = { path = "../pd_client", default-features = false }
+prometheus = { version = "0.13", features = ["nightly"] }
+protobuf = { version = "2.8", features = ["bytes"] }
+raft = { version = "0.7.0", default-features = false, features = ["protobuf-codec"] }
+raft_log_engine = { path = "../raft_log_engine", default-features = false }
+raftstore = { path = "../raftstore", default-features = false }
+rand = "0.8"
+resolved_ts = { path = "../../components/resolved_ts", default-features = false }
+resource_metering = { path = "../resource_metering" }
+security = { path = "../security", default-features = false }
+serde = "1.0"
+serde_derive = "1.0"
+serde_ignored = "0.1"
+serde_json = "1.0"
+slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
+slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
+tempfile = "3.0"
+tikv = { path = "../..", default-features = false }
+tikv_alloc = { path = "../tikv_alloc" }
+tikv_util = { path = "../tikv_util", default-features = false }
+tokio = { version = "1.5", features = ["rt-multi-thread"] }
+toml = "0.5"
+txn_types = { path = "../txn_types", default-features = false }
+server = { path = "../server", default-features = false }
+yatp = { git = "https://github.com/tikv/yatp.git", branch = "master" }
+
+[target.'cfg(unix)'.dependencies]
+signal = "0.6"
