+++ mock-engine-store/Cargo.toml
@@ -25,6 +25,7 @@ kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = f
 protobuf = "2.8.0"
 raftstore = { path = "../components/raftstore", default-features = false }
 server = { path = "../components/server" }
+proxy_server = { path = "../components/proxy_server" }
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
 tikv_util = { path = "../components/tikv_util", default-features = false }
