+++ new-mock-engine-store/Cargo.toml
@@ -49,3 +49,4 @@ test_raftstore = { path = "../components/test_raftstore" }
 test_util = { path = "../components/test_util", default-features = false }
 tikv = { path = "../.", default-features = false }
 tikv_util = { path = "../components/tikv_util", default-features = false }
+proxy_server = { path = "../components/proxy_server" }
