+++ raftstore-proxy/src/lib.rs
@@ -4,7 +4,7 @@ use std::os::raw::{c_char, c_int};
 /// Print version infomatin to std output.
 #[no_mangle]
 pub unsafe extern "C" fn print_raftstore_proxy_version() {
-    server::print_proxy_version();
+    proxy_server::print_proxy_version();
 }
 
 /// # Safety
@@ -15,5 +15,5 @@ pub unsafe extern "C" fn run_raftstore_proxy_ffi(
     argv: *const *const c_char,
     helper: *const u8,
 ) {
-    server::run_proxy(argc, argv, helper);
+    proxy_server::run_proxy(argc, argv, helper);
 }
