+++ tests/Cargo.toml
@@ -81,6 +81,7 @@ raft_log_engine = { path = "../components/raft_log_engine", default-features = f
 raftstore = { path = "../components/raftstore", default-features = false }
 rand = "0.8.3"
 server = { path = "../components/server", default-features = false }
+proxy_server = { path = "../components/proxy_server", default-features = false }
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
 tempfile = "3.0"
